Slight cleanup after portals

+++ src/Generating/CompoGen.cpp
@@ -638,7 +638,7 @@ void cCompoGenNether::ComposeTerrain(cChunkDesc & a_ChunkDesc)
 			CeilingDisguise = -CeilingDisguise;
 		}
 
-		int CeilingDisguiseHeight = Height - 2 - CeilingDisguise * 3;
+		int CeilingDisguiseHeight = Height - 2 - (int)CeilingDisguise * 3;
 
 		for (int y = Height - 1; y > CeilingDisguiseHeight; y--)
 		{
